Refactor the code

+++ src/main/kotlin/com/springkotlin/springsplit/dto/SplitDTO.kt
@@ -1,9 +1,7 @@
 package com.springkotlin.springsplit.dto
 
-data class SplitDTO (
-        val amount:Int,
-
-        val splitList:List<String>,
-        val troopName:String
-
+data class SplitDTO(
+        val amount: Float,
+        val splitList: List<String>,
+        val troopName: String,
 )
